requirements.txt
# Core ML Dependencies
tensorflow>=2.15.0
numpy>=1.26.0
matplotlib>=3.7.1
scikit-learn>=1.3.0

# Web Application
Flask>=3.0.0
Werkzeug>=3.0.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
